MDTZ-977 Add schema validation for incoming requests
