+++ src/infrastructure/figma/index.ts
@@ -1,4 +1,3 @@
 export * from './figma-auth-service';
 export * from './figma-service';
 export * from './errors';
-export * from './schemas';
