+++ src/usecases/index.ts
@@ -4,4 +4,6 @@ export * from './uninstalled-use-case';
 export * from './check-user-figma-auth-use-case';
 export * from './associate-entity-use-case';
 export * from './disassociate-entity-use-case';
-export * from './handle-figma-file-change-event-use-case';
+export * from './handle-figma-file-update-event-use-case';
+export * from './types';
+export * from './schemas';
