+++ src/web/middleware/errors.ts
@@ -1,0 +1,1 @@
+export class UnauthorizedError extends Error {}
