+++ src/web/middleware/extract-user-id-from-headers-middleware.ts
@@ -1,0 +1,20 @@
+import type { NextFunction, Request, Response } from 'express';
+
+import { UnauthorizedError } from './errors';
+
+import { ensureString } from '../../common/stringUtils';
+
+export const extractUserIdFromHeadersMiddleware = (
+	req: Request,
+	res: Response<unknown, { atlassianUserId: string }>,
+	next: NextFunction,
+): void => {
+	try {
+		const userId = ensureString(req.headers['user-id']);
+		res.locals.atlassianUserId = userId;
+	} catch (e: unknown) {
+		return next(new UnauthorizedError('Missing or invalid User-Id header'));
+	}
+
+	next();
+};
