+++ src/web/middleware/index.ts
@@ -1,3 +1,4 @@
 export * from './http-logger-middleware';
 export * from './auth-header-jwt-middleware';
 export * from './error-handler-middleware';
+export * from './extract-user-id-from-headers-middleware';
