+++ src/web/routes/entities/types.ts
@@ -1,4 +1,4 @@
-import type { Response } from 'express';
+import type { Request, Response } from 'express';
 
 import type {
 	AtlassianDesign,
@@ -9,22 +9,45 @@ import type {
 	DisassociateEntityUseCaseParams,
 } from '../../../usecases';
 
-export type AssociateEntityRequestParams = Omit<
+export type EntitiesResponseBody = { design: AtlassianDesign } | string;
+
+export type EntitiesRequestLocals = {
+	connectInstallation: ConnectInstallation;
+	atlassianUserId: string;
+};
+
+export type AssociateEntityRequestBody = Omit<
 	AssociateEntityUseCaseParams,
 	'atlassianUserId' | 'connectInstallation'
 >;
 
-export type AssociateEntityResponse = Response<
-	{ design: AtlassianDesign } | string,
-	{ connectInstallation: ConnectInstallation }
+export type AssociateEntityRequest = Request<
+	Record<string, never>,
+	EntitiesResponseBody,
+	AssociateEntityRequestBody,
+	Record<string, never>,
+	EntitiesRequestLocals
 >;
 
-export type DisassociateEntityRequestParams = Omit<
+export type AssociateEntityResponse = Response<
+	{ design: AtlassianDesign } | string,
+	EntitiesRequestLocals
+>;
+
+export type DisassociateEntityRequestBody = Omit<
 	DisassociateEntityUseCaseParams,
 	'atlassianUserId' | 'connectInstallation'
 >;
 
+export type DisassociateEntityRequest = Request<
+	Record<string, never>,
+	EntitiesResponseBody,
+	DisassociateEntityRequestBody,
+	Record<string, never>,
+	EntitiesRequestLocals
+>;
+
 export type DisassociateEntityResponse = Response<
-	{ design: AtlassianDesign } | string,
-	{ connectInstallation: ConnectInstallation }
+	EntitiesResponseBody,
+	EntitiesRequestLocals
 >;
