--- src/web/routes/types.ts
@@ -1,5 +1,0 @@
-import type { Request } from 'express';
-
-export interface TypedRequest<T> extends Request {
-	body: T;
-}
